Move Pagination component into UI::Pagination namespace
