+++ app/components/admin/pagination_with_count/component.html.erb
@@ -25,7 +25,7 @@
 
   <% unless @skip_pagination || @pagy.blank? %>
     <div class="tw:flex tw:flex-wrap tw:justify-end tw:gap-4 tw:md:flex-nowrap tw:md:text-right <%= (@skip_total ? "col-12" : "col-md-7") %>">
-      <%= render(Pagination::Component.new(pagy: @pagy, page_params: @params)) %>
+      <%= render(UI::Pagination::Component.new(pagy: @pagy, page_params: @params)) %>
 
       <%= select_tag :per_page_select, options_for_select(per_pages.map { |i| ["#{i} / page", i] }, selected: @per_page), {id: per_page_select_id, class: "form-control d-inline-block", style: "margin-top: -6px; max-width: 8rem;"} %>
     </div>
