+++ app/components/org/impound_records_index/component.html.erb
@@ -152,5 +152,5 @@
 <% end %>
 
 <div class="pt-4 pb-4">
-  <%= render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
+  <%= render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
 </div>
