+++ app/views/my_accounts/messages/index.html.erb
@@ -9,5 +9,5 @@
     <% end %>
   <% end %>
 
-  <%= render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
+  <%= render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
 </div>
