+++ app/views/my_accounts/show.html.haml
@@ -115,7 +115,7 @@
                         = link_to t(".ignore"), user_alert_path(id: alert.id, alert_action: "dismiss"), method: :patch, action: "update"
 
             .pt-4.pb-4
-              = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+              = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
           - else
             %h2= t(".you_have_no_registered_bikes")
         .tab-pane#locks_tab_content{ class: ('active' if @locks_active_tab) }
