+++ app/views/org_public/impounded_bikes/index.html.haml
@@ -57,7 +57,7 @@
               = impound_record.formatted_address_string(visible_attribute: :street)
 
   .pt-4.pb-4
-    = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+    = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
 
 -# include javascript because of period select
 <script src="/vendored_assets/application.js"></script>
