+++ app/views/organized/bikes/_list.html.haml
@@ -76,4 +76,4 @@
                 method: "PUT"
 
 .pt-4.pb-4
-  = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+  = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
