+++ app/views/organized/bikes/recoveries.html.haml
@@ -68,4 +68,4 @@
                 = bike.owner_email
 
 .pt-4.pb-4
-  = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+  = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
