+++ app/views/organized/bulk_imports/index.html.erb
@@ -136,5 +136,5 @@
 </table>
 
 <div class="pb-4 pt-4">
-  <%= render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
+  <%= render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
 </div>
