+++ app/views/organized/exports/index.html.erb
@@ -99,6 +99,6 @@
   </div>
 
   <div class="pb-4 pt-4">
-    <%= render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
+    <%= render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
   </div>
 </div>
