+++ app/views/organized/model_audits/index.html.haml
@@ -70,6 +70,6 @@
             update
 
 .pt-4.pb-4
-  = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+  = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
 
 = render partial: "/organized/model_audits/submit_model_attestation"
