+++ app/views/organized/users/index.html.haml
@@ -31,7 +31,7 @@
 .row.mb-4.mt-4
   .col-md-6
     .pt-4.pb-4
-      = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+      = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
   - if @show_user_search
     .col-md-6.text-right{ style: "margin-top: 0.25rem;" }
       = form_tag organization_users_path(organization_id: current_organization.to_param), method: :get, class: "form-inline" do
