+++ app/views/stolen_bike_listings/index.html.haml
@@ -127,7 +127,7 @@
     .col-md-4.mt-4
       .ad-block.ad-binx.ad468x60
     .col-md-8.col-lg-4.pt-4.pb-4
-      = render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
+      = render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg))
 
 
   - unless @stolen_bike_listings.present?
