+++ app/views/welcome/recovery_stories.html.erb
@@ -21,7 +21,7 @@
     </div>
 
     <div class="pt-4 pb-4">
-      <%= render(Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
+      <%= render(UI::Pagination::Component.new(pagy: @pagy, page_params: params, size: :lg)) %>
     </div>
   </div>
 </div>
